:construction: usermode highjacking

+++ WinLib/WinLibKernel/MMapDll.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <ntifs.h>
 #include "PEFile.h"
 #include "ntos.h"
 
@@ -16,7 +17,7 @@ namespace WinLibKernel {
 
 				MMapperDll(PEFile *peFile);
 
-				STATUS map(PEPROCESS process, PVOID originalEntryPoint, PVOID targetBase, DWORD targetSize);
+				STATUS map(PEPROCESS process, PVOID originalEntryPoint, PVOID targetBase, DWORD64 targetSize);
 			private:
 				PEFile * peFile;
 				PVOID payload;
@@ -25,7 +26,7 @@ namespace WinLibKernel {
 				bool mapSections();
 				bool baseRelocation(PVOID targetBase);
 				bool fixImports();
-				bool writeToProcess(PEPROCESS process, PVOID targetBase, DWORD targetSize);
+				bool writeToProcess(PEPROCESS process, PVOID targetBase, DWORD64 targetSize);
 				bool patchEntryPoint(PVOID originalEntryPoint);
 
 				PIMAGE_BASE_RELOCATION processRelocation(ULONG_PTR address, ULONG count, PUSHORT typeOffset, LONGLONG delta);
